+++ ui-components/src/page_components/listing/UnitTables.tsx
@@ -16,13 +16,9 @@ const formatRange = (range: MinMax, ordinalize?: boolean) => {
   }
 
   if (min == max) {
-    return <>{min}</>
+    return min
   } else {
-    return (
-      <>
-        {min} - {max}
-      </>
-    )
+    return `${min} - ${max}`
   }
 }
 
@@ -80,22 +76,16 @@ const UnitTables = (props: UnitTablesProps) => {
 
         let areaRangeSection
         if (unitSummary.areaRange?.min || unitSummary.areaRange?.max) {
-          areaRangeSection = (
-            <>
-              ,&nbsp;{formatRange(unitSummary.areaRange)} {t("t.squareFeet")}
-            </>
-          )
+          areaRangeSection = `, ${formatRange(unitSummary.areaRange)} ${t("t.squareFeet")}`
         }
 
         if (unitSummary.floorRange && unitSummary.floorRange.min) {
-          floorSection = (
-            <>
-              ,&nbsp;{formatRange(unitSummary.floorRange, true)}{" "}
-              {unitSummary.floorRange.max > unitSummary.floorRange.min
-                ? t("t.floors")
-                : t("t.floor")}
-            </>
-          )
+          floorSection = `, ${formatRange(unitSummary.floorRange, true)} 
+              ${
+                unitSummary.floorRange.max > unitSummary.floorRange.min
+                  ? t("t.floors")
+                  : t("t.floor")
+              }`
         }
 
         return (